Remove extra spaces

+++ portfolio/src/main/java/com/google/sps/servlets/DataServlet.java
@@ -57,8 +57,7 @@ public class DataServlet extends HttpServlet {
       response.sendError(HttpServletResponse.SC_BAD_REQUEST, ERR_MSG);
       return;
     }
-    Query query = new Query("Comment")
-                    .addSort("like", SortDirection.DESCENDING);
+    Query query = new Query("Comment").addSort("like", SortDirection.DESCENDING);
     PreparedQuery pq = this.datastore.prepare(query);
 
     List<String> comments = pq.asList(FetchOptions.Builder.withLimit(limit))
@@ -89,7 +88,7 @@ public class DataServlet extends HttpServlet {
     Entity commentEntity = new Entity("Comment");
     commentEntity.setProperty("content", newComment);
     commentEntity.setProperty("timestamp", timestamp);
-    commentEntity.setProperty("like", (long)0);
+    commentEntity.setProperty("like", 0);
     this.datastore.put(commentEntity);
   }
